requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Microsoft Agent Framework
agent-framework-core==1.0.0rc3
agent-framework-claude==1.0.0b260225
agent-framework-azure-ai==1.0.0rc3
azure-ai-agentserver-agentframework==1.0.0b17
claude-agent-sdk==0.1.56
opentelemetry-sdk>=1.33.1,<2
opentelemetry-exporter-otlp-proto-http>=1.33.1,<2

# Agent Governance Toolkit
agent-governance-toolkit[full]

# Utilities
pyyaml>=6.0
jsonschema>=4.0.0
colorama>=0.4.6
tabulate>=0.9.0
